Split form structure specific components and form prop specific components
