+++ mathesar_ui/src/systems/data-forms/add-edit-modal/index.ts
@@ -1,0 +1,1 @@
+export { default as AddEditDataFormModal } from './AddEditDataFormModal.svelte';
