+++ mathesar_ui/src/systems/data-forms/form-maker/config-panel/DataFormConfigPanel.svelte
