+++ mathesar_ui/src/systems/data-forms/form-maker/config-panel/form-config/SubmissionSettings.svelte
